Print dvc version info for debugging
fix #4095
1. seperate info from dvc version.
2. print dvc version info when some unexpected error occured.
